ui: merge image playground into image gen page.
