+++ frontend/src/i18n/locales/zh.ts
@@ -47,7 +47,6 @@ export default {
       "useEmbed": "Embedding",
       "useImageGen": "Image Gen",
       "useTeam": "Team",
-      "playground": "Playground",
       "apiKeys": "API 密钥",
       "oauth": "OAuth 凭证",
       "credential": "凭证",
